ci: toolchain pin → 1.98.1 + cargo-semver-checks 0.50.0 明示 pin
1.98.1 新規 lint 修正: sort / fill / 冗長 cast / saturating_sub.max(0) / zip loop (演算順序不変) / checked_div / doc indent 21 件 + pub 1.0 API 2 fn の too_many_arguments は semver 理由付き allow

真因: 1.92 pin のまま放置 (月次 bump 方針未実施) → semver-checks@latest (MSRV 1.93) が赤 (残 8 repo のうち 6、Streaming-Protocol / View は別 session 作業中で skip)

+++ src/math_util.rs
@@ -97,7 +97,7 @@ pub fn cbrt_fix(n: Fix128) -> Fix128 {
     // Initial guess: use bit-position estimate to seed Newton.
     let mut x = if n >= Fix128::ONE {
         // n >= 1 → cbrt ≥ 1 → seed with hi/2 (rough)
-        Fix128::from_int(1 + (n.hi.max(1) as i64) / 2)
+        Fix128::from_int(1 + n.hi.max(1) / 2)
     } else {
         // n < 1 → seed slightly below 1
         Fix128::from_ratio(5, 10)
